Split service process control out to service manager class
