+++ rb/lib/selenium/webdriver/common.rb
@@ -23,6 +23,7 @@ require 'selenium/webdriver/common/proxy'
 require 'selenium/webdriver/common/log_entry'
 require 'selenium/webdriver/common/file_reaper'
 require 'selenium/webdriver/common/service'
+require 'selenium/webdriver/common/service_manager'
 require 'selenium/webdriver/common/socket_lock'
 require 'selenium/webdriver/common/socket_poller'
 require 'selenium/webdriver/common/port_prober'
