+++ rb/lib/selenium/webdriver/common/service.rb
@@ -21,14 +21,10 @@ module Selenium
   module WebDriver
     #
     # Base class implementing default behavior of service object,
-    # responsible for starting and stopping driver implementations.
+    # responsible for storing a service manager configuration.
     #
 
     class Service
-      START_TIMEOUT = 20
-      SOCKET_LOCK_TIMEOUT = 45
-      STOP_TIMEOUT = 20
-
       class << self
         attr_reader :driver_path
 
@@ -81,7 +77,7 @@ module Selenium
       end
 
       attr_accessor :host
-      attr_reader :executable_path
+      attr_reader :executable_path, :port, :extra_args
 
       #
       # End users should use a class method for the desired driver, rather than using this directly.
@@ -103,31 +99,20 @@ module Selenium
         raise Error::WebDriverError, "invalid port: #{@port}" if @port < 1
       end
 
-      def start
-        raise "already started: #{uri.inspect} #{@executable_path.inspect}" if process_running?
-
-        Platform.exit_hook(&method(:stop)) # make sure we don't leave the server running
-
-        socket_lock.locked do
-          find_free_port
-          start_process
-          connect_until_stable
-        end
+      def launch
+        sm = ServiceManager.new(self)
+        sm.start
+        sm
       end
 
-      def stop
-        return unless self.class::SHUTDOWN_SUPPORTED
-
-        stop_server
-        @process.poll_for_exit STOP_TIMEOUT
-      rescue ChildProcess::TimeoutError
-        nil # noop
-      ensure
-        stop_process
+      def shutdown_supported
+        self.class::SHUTDOWN_SUPPORTED
       end
 
-      def uri
-        @uri ||= URI.parse("http://#{@host}:#{@port}")
+      protected
+
+      def extract_service_args(driver_opts)
+        driver_opts.key?(:args) ? driver_opts.delete(:args) : []
       end
 
       private
@@ -141,85 +126,6 @@ module Selenium
         Platform.assert_executable path
         path
       end
-
-      def build_process(*command)
-        WebDriver.logger.debug("Executing Process #{command}")
-        @process = ChildProcess.build(*command)
-        if WebDriver.logger.debug?
-          @process.io.stdout = @process.io.stderr = WebDriver.logger.io
-        elsif Platform.jruby?
-          # Apparently we need to read the output of drivers on JRuby.
-          @process.io.stdout = @process.io.stderr = File.new(Platform.null_device, 'w')
-        end
-
-        @process
-      end
-
-      def connect_to_server
-        Net::HTTP.start(@host, @port) do |http|
-          http.open_timeout = STOP_TIMEOUT / 2
-          http.read_timeout = STOP_TIMEOUT / 2
-
-          yield http
-        end
-      end
-
-      def find_free_port
-        @port = PortProber.above(@port)
-      end
-
-      def start_process
-        @process = build_process(@executable_path, "--port=#{@port}", *@extra_args)
-        # Note: this is a bug only in Windows 7
-        @process.leader = true unless Platform.windows?
-        @process.start
-      end
-
-      def stop_process
-        return if process_exited?
-
-        @process.stop STOP_TIMEOUT
-        @process.io.stdout.close if Platform.jruby? && !WebDriver.logger.debug?
-      end
-
-      def stop_server
-        return if process_exited?
-
-        connect_to_server do |http|
-          headers = WebDriver::Remote::Http::Common::DEFAULT_HEADERS.dup
-          http.get('/shutdown', headers)
-        end
-      end
-
-      def process_running?
-        defined?(@process) && @process&.alive?
-      end
-
-      def process_exited?
-        @process.nil? || @process.exited?
-      end
-
-      def connect_until_stable
-        socket_poller = SocketPoller.new @host, @port, START_TIMEOUT
-        return if socket_poller.connected?
-
-        raise Error::WebDriverError, cannot_connect_error_text
-      end
-
-      def cannot_connect_error_text
-        "unable to connect to #{self.class::EXECUTABLE} #{@host}:#{@port}"
-      end
-
-      def socket_lock
-        @socket_lock ||= SocketLock.new(@port - 1, SOCKET_LOCK_TIMEOUT)
-      end
-
-      protected
-
-      def extract_service_args(driver_opts)
-        driver_opts.key?(:args) ? driver_opts.delete(:args) : []
-      end
-
     end # Service
   end # WebDriver
 end # Selenium
