+++ rb/spec/unit/selenium/webdriver/edge/driver_spec.rb
@@ -24,14 +24,13 @@ module Selenium
     module EdgeHtml
       describe Driver do
         let(:resp)    { {'value' => {'sessionId' => 'foo', 'capabilities' => Remote::Capabilities.edge.as_json}} }
-        let(:service) { instance_double(Service, start: nil, uri: 'http://example.com') }
+        let(:service) { instance_double(Service) }
         let(:caps)    { Remote::Capabilities.edge }
         let(:http)    { instance_double(Remote::Http::Default, call: resp).as_null_object }
 
         before do
           allow(Remote::Capabilities).to receive(:internet_explorer).and_return(caps)
-          allow(Service).to receive(:binary_path).and_return('/foo')
-          allow(Service).to receive(:new).and_return(service)
+          allow(Service).to receive_messages(binary_path: '/foo', new: service)
         end
 
         it 'accepts server URL' do
