+++ rb/spec/unit/selenium/webdriver/ie/service_spec.rb
@@ -36,7 +36,7 @@ module Selenium
 
           expect(service.executable_path).to include IE::Service::EXECUTABLE
           expected_port = IE::Service::DEFAULT_PORT
-          expect(service.uri.to_s).to eq "http://#{Platform.localhost}:#{expected_port}"
+          expect(service.port).to eq expected_port
         end
 
         it 'uses provided path and port' do
@@ -46,7 +46,8 @@ module Selenium
           service = Service.ie(path: path, port: port)
 
           expect(service.executable_path).to eq path
-          expect(service.uri.to_s).to eq "http://#{Platform.localhost}:#{port}"
+          expect(service.port).to eq port
+          expect(service.host).to eq Platform.localhost
         end
 
         it 'allows #driver_path= with String value' do
@@ -89,7 +90,7 @@ module Selenium
 
           service = Service.ie
 
-          expect(service.instance_variable_get('@extra_args')).to be_empty
+          expect(service.extra_args).to be_empty
         end
 
         it 'uses provided args' do
@@ -97,7 +98,7 @@ module Selenium
 
           service = Service.ie(args: ['--foo', '--bar'])
 
-          expect(service.instance_variable_get('@extra_args')).to eq ['--foo', '--bar']
+          expect(service.extra_args).to eq ['--foo', '--bar']
         end
 
         # This is deprecated behavior
@@ -107,21 +108,24 @@ module Selenium
           service = Service.ie(args: {log_file: '/path/to/log',
                                       silent: true})
 
-          expect(service.instance_variable_get('@extra_args')).to eq ['--log-file=/path/to/log', '--silent']
+          expect(service.extra_args).to eq ['--log-file=/path/to/log', '--silent']
         end
       end
     end
 
     module IE
       describe Driver do
-        let(:service) { instance_double(Service, start: true, uri: 'http://example.com') }
+        let(:service) { instance_double(Service, launch: service_manager) }
         let(:bridge) { instance_double(Remote::Bridge, quit: nil, create_session: {}) }
+        let(:service_manager) { instance_double(ServiceManager, uri: 'http://example.com') }
 
         before do
           allow(Remote::Bridge).to receive(:new).and_return(bridge)
+          allow(ServiceManager).to receive(:new).and_return(service_manager)
         end
 
         it 'is not created when :url is provided' do
+          expect(ServiceManager).not_to receive(:new)
           expect(Service).not_to receive(:new)
 
           described_class.new(url: 'http://example.com:4321')
@@ -136,9 +140,9 @@ module Selenium
         it 'accepts :driver_path but throws deprecation notice' do
           driver_path = '/path/to/driver'
 
-          expect(Service).to receive(:new).with(path: driver_path,
-                                                port: nil,
-                                                args: nil).and_return(service)
+          allow(Service).to receive(:new).with(path: driver_path,
+                                               port: nil,
+                                               args: nil).and_return(service)
 
           expect {
             described_class.new(driver_path: driver_path)
@@ -148,9 +152,9 @@ module Selenium
         it 'accepts :port but throws deprecation notice' do
           driver_port = 1234
 
-          expect(Service).to receive(:new).with(path: nil,
-                                                port: driver_port,
-                                                args: nil).and_return(service)
+          allow(Service).to receive(:new).with(path: nil,
+                                               port: driver_port,
+                                               args: nil).and_return(service)
 
           expect {
             described_class.new(port: driver_port)
@@ -161,9 +165,9 @@ module Selenium
           driver_opts = {foo: 'bar',
                          bar: ['--foo', '--bar']}
 
-          expect(Service).to receive(:new).with(path: nil,
-                                                port: nil,
-                                                args: driver_opts).and_return(service)
+          allow(Service).to receive(:new).with(path: nil,
+                                               port: nil,
+                                               args: driver_opts).and_return(service)
 
           expect {
             described_class.new(driver_opts: driver_opts)
